Change type for Pd and Type

+++ ARMDServer/Request.cs
@@ -7,8 +7,8 @@ namespace ARMDServer
     public struct Request
     {
         public readonly uint Identifier { get; }
-        public readonly long Pd { get; }
-        public readonly long Type { get; }
+        public readonly ushort Pd { get; }
+        public readonly ushort Type { get; }
         public readonly BinaryDateTime CncTime { get; }
 
         public bool IsValid => Identifier == 0x42535253 && Pd == 1 && Type == 1;
